requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-cors-headers>=4.3
python-dotenv>=1.0
Pillow>=10.0
google-genai>=0.3.0
